+++ public/Publish-AdfV2UsingArm.ps1
@@ -124,7 +124,7 @@ function Publish-AdfV2UsingArm {
         }
 
         if ($null -eq $targetAdf) {
-            Write-Host "The process is exiting the function. Do fix the issue and run again."
+            Write-Host "ADFT0032: The process is exiting the function. Do fix the issue and run again."
             return 
         }
     }